Feature: Added Logout Routing
